+++ supervisor/create_windows.go
@@ -1,0 +1,8 @@
+package supervisor
+
+type platformStartTask struct {
+}
+
+// Checkpoint not supported on Windows
+func (task *startTask) setTaskCheckpoint(t *StartTask) {
+}
